【spring事物分析】

+++ src/main/java/com/nivelle/guide/springboot/mapper/ActivityPvMapper.java
@@ -23,7 +23,7 @@ public interface ActivityPvMapper {
 
     int update(ActivityPvEntity activityPvEntity);
 
-    ActivityPvEntity getActivityById(Integer id);
+    ActivityPvEntity getActivityById(long id);
 
 
 }
